+++ src/ex5/server.js
@@ -1,7 +1,5 @@
-// Express boilerplate, hosting the `dist` file, connecting to the routes
 const fs = require("fs");
 const express = require("express");
-const bodyParser = require("body-parser");
 
 require("express-async-errors");
 
@@ -14,14 +12,6 @@ const server = express();
 
 server.use([logger, express.json(), express.static("dist")]);
 server.use("/item", itemRouter);
-
-server.use(bodyParser.json()); // to support JSON-encoded bodies
-server.use(
-  bodyParser.urlencoded({
-    // to support URL-encoded bodies
-    extended: true,
-  })
-);
 
 server.get("/", (req, res) => {
   res.status(200).json({
